Change location actions to be in locations controller
