+++ src/Promenade.Web/Views/Locations/LocationDetails.cshtml
@@ -1,4 +1,4 @@
-﻿@model Ocuda.Promenade.Controllers.ViewModels.LocationViewModel
+﻿@model Ocuda.Promenade.Controllers.ViewModels.Locations.LocationViewModel
 @{
     ViewData["Title"] = "LocationDetails";
 }
