+++ src/Promenade.Web/Views/Locations/Locations.cshtml
@@ -1,4 +1,4 @@
-﻿@model Ocuda.Promenade.Controllers.ViewModels.LocationViewModel
+﻿@model Ocuda.Promenade.Controllers.ViewModels.Locations.LocationViewModel
 @{
     ViewData["Title"] = "Locations";
 }
